feat: add description for VIP-256
