+++ simulations/vip-256/abi/beacon.json
