+++ simulations/vip-256/abi/resilientOracle.json
